Replace travis & python-publish with full CI/CD GitHub Actions setup
When releasing, you simply need to tag a commit with the tag starting with `v`, e.g., `v1.4.0`. It will then automatically build a wheel, make a GitHub release, attach the wheel to the GitHub release, then publish the Wheel and Srcdist to PyPI.

A side benefit is that it also automatically creates a generic changelog based on the commits since the previous release.

The CI workflow does the same as what the travis workflow did.