Replace use of Embedded with nested "data" field
This is by no means ideal. It would be much preferred to be embedded, but the maintainer of construct has stated there is no workaround and nesting is the only option.

This means when building boxes we must now nest the data in `data` container. This also means grabbing a fields value must now specify `.data` first, e.g., `tenc_box.data.is_encrypted`. See changes with the tests. Not ideal at all.